Implement proper cascade of settings
Until now, all members of the Settings object were determined by the
command-line parser, which used hardcoded default. This means that we
never included any settings whatsoever from the environment or the
configuration file.

Fix this by changing the command-line parsing so that only arguments
that are actually _passed_ on the command line are passed on to the
Settings constructor. This will cause settings from the environment or
the configuration file to be applied, since they are no longer always
overridden by the command line.