Store Settings objects inside Analysis (and include in JSON output)
Until now, we have stored the set of requested actions inside the
Analysis object, and hidden it from the JSON output.

Now that we have a structured Settings object that encapsulates all of
our configuration (including the requested actions), it makes sense to
store that in the Analysis object instead.

Also expose this object in the JSON output. This makes the JSON output
considerably more verbose, but also makes it much easier to debug when
there might be retroactive questions about what FawltyDeps was asked to
do.

In order to include the entire Settings object in the JSON output and
successfully test this, we need the JSON serialization of our Settings
object to be stable. Fortunately it's fairly easy to customize the
pydantic JSON encoder to use sorted() instead of list() when serializing
(frozen)set objects.