Add a new experiment (with a `requirements` field) to peppy and algorithms
The new `requirements` field is a list of requirements (with optional
specifications), in th same format expected by a requirements.txt file.
These requirements will be passed to the `pip` command to install packages in a
specific virtual environment.

Note that we did not define the `imports` and `declared_deps` fields in this
experiment, as these are already defined (and therefore tested) in another
experiment. In defining the requirements of a virtual environment in which
FawltyDeps will run, we only expect the `undeclared_deps` and `unused_deps` to
be affected. To expect changes in other fields of `Analysis`, other args must be
changed (e.g. `--deps` or `--code`).

Co-authored-by: Johan Herland <johan.herland@tweag.io>