types: Remove mapping details from DeclaredDependency
There is no longer any need to store the import names and mapping enum
details _inside_ the DeclaredDependency class. If we do want these
details in the (e.g. JSON) output of FawltyDeps, we can instead add the
corresponding Package objects back into the UnusedDependency objects.

For now, though, simply remove the .import_names and .mapping members
from the DeclaredDependency class, along with associated code and test
cases.

This helps cut down on the duplication inside the internal Analysis data
structure, which is also reflected in the JSON output.