types: Reduce duplication by changing .references into List[Location]
Until now we had:

    class UndeclaredDependency:
        name: str
	references: List[ParsedImport]

    class UnusedDependency:
        name: str
	references: List[DeclaredDependency]

but both ParsedImport and DeclaredDependency contained their own .name
which was identical to the corresponding .name above. This duplication
is unnecessary, and we can avoid it by storing the .source member of the
latter objects in the .references lists of the respective former
objects. In other words:

    class UndeclaredDependency:
        name: str
	references: List[Location]

    class UnusedDependency:
        name: str
	references: List[Location]

The simpler structure is nicely reflected in some shorter test cases.