main: Add information on resolved dependencies to Analysis object
When we removed .import_names and .mapping from DeclaredDependency, we
also remove any information in our JSON output that could help debug the
dependency-to-import-names mapping that FD performs.

This commits adds that information back in a more useful format. Instead
of duplicating it across many DelaredDependency objects, we now add a
dedicated field to the top-level Analysis object called .resolved_deps.

Analysis.resolved_deps is a mapping from dependency names (as they occur
in the --deps files we have parsed) to the corresponding Package object,
which contains information on what import names that FawltyDeps found,
and which mapping it used to find those import names (currently either
via the local environment, or by falling back to the identity mapping).