Package: Restructure .mappings member to keep track of import names
As Maria suggests in [1], knowing which import name came from which
mapping can be very useful when debugging. We achieve this by combining
.mappings and .import_names into a dict that associates each mapping
with the import names it contributes.

However, although now redundant, we still want to hang on to the
flattened .import_names set for checking if each declared dep is used.

Therefore we assign the new dict to .mappings and keep the redundant
.import_names, except that we hide it from our JSON output.

[1]: https://github.com/tweag/FawltyDeps/pull/193#discussion_r1123805564