shell.nix: Fix venv creation on python3.7
For Reasons(tm), python3.7 does not work 100% when put in the same
nix-shell as more recent Python versions. Symptoms include that venv
creation fails with:

  ModuleNotFoundError: No module named '_sysconfigdata__linux_x86_64-linux-gnu'

One workaround is to simply list 'python37' _after_ the other 'python*'s
in the 'buildInputs' list in shell.nix. However, that is too magic to be
considered a real workaround.

A slightly less magic workaround is suggested in [1]: unset a couple of
environment variables that seem to trip up Python 3.7.

[1]: https://github.com/NixOS/nixpkgs/issues/88711