test_compare_imports_to_dependencies: Refactor and add tests
We will soon remove compare_imports_to_dependencies(), but we have a lot
of parametrized tests for this function that we want to retain.

Turn each test vector into a FDTestVector instance to make the
definition of each vector more readable, and then use the same test data
to test resolve_dependencies(), calculate_undeclared(), and
calculate_unused() in addition to compare_imports_to_dependencies().

Based on Maria's great idea in [1].

[1]: https://github.com/tweag/FawltyDeps/pull/193#discussion_r1124164566