test_sample_projects: Add optional checks and more verbose debug info
Add some print() statements while testing each sample project:
 - Print the name and description of the project
 - Print a line before each check that we perform

These prints are only visible when the test fails, and helps us more
easily see which project and which check is failing.

Also add optional verification of analysis.imports and .declared_deps.
If the "imports = [ ... ]" or "declared_deps = [ ... ]" exist in the
expected.toml, then we will verify them in addition to the verification
of .undeclared_deps and .unused_deps. Otherwise, we only check the
latter two (which are mandatory).