sample_projects: Add example from announcement blog post
In our announcement blog post we introduce a contrived example to
demonstrate the benefits of using FawltyDeps in a simple Python project.
We need somewhere to point our readers where they can study this example
in more detail.

At the same time, we want to make sure that FawltyDeps behaves
predictably in this small example project. Hence it is a perfect
candidate for our sample_projects framework.