test_real_projects: Use context manager instead of autouse fixure
As part of each real_projects experiment, we need to install FD in
editable mode just before running the test, and uninstall it afterwards.
Until now this has been accomplished with an autouse fixture, but those
are a little too magic, IMHO, and it's clearer what is happening if we
instead use an explicit context manager to both do the setup/teardown of
the venv_dir-with-editable-install, as well as as the communication of
the actual venv_dir into the test case.

Fixes: #199