resolve_dependencies(): Log INFO message when we fall back to id mapping
We want the user to be able to see when FD has fallen back to using the
inferior identity mapping. Before this, the user would have to use the
`--json` option and look at the contents of `resolved_deps`.

This adds an INFO-level log message once for each package where we
resort to the identity mapping. This is now available to the user by
passing the `--verbose` option.

Adjust tests accordingly. Note: The test_cmdline are getting out of hand
and needs some thorough refactoring to not be a stumbling block for
pretty much every code change we make in the future...