test_real_projects: Verify that all requirements are present
LocalPackageLookup recently learned how to peek into another venv and
find its installed packages. Use this feature in test_real_projects to
verify that the requirements declared for the current Experiment are in
fact present, before we proceed with testing the expected result from
running FD on this Experiment.

This is a relatively cheap way, both to verify that our cached venvs are
OK, and to test that LocalPackageLookup in a different venv works
correctly.