pyproject.toml: Use latest importlib_metadata across all Python versions
We're about to start relying on internal functions from
importlib_metadata. Although these are available via the stdlib module
importlib.metadata in Python v3.11, we are likely keep relying on
upcoming bugfixes from the importlib_metadata project, and it is
probably safer for us to _pin_ this 3rd-party dependency, so that we can
avoid breakage if/when its internal functions change.