packages.LocalPackageLookup: Remove redundant sanity check
When we added the venv_path argument to LocalPackageLookup, we retained
a sanity check for the default venv_path=None case, where we compared
the new method of building .packages against the old method method
(using packages_distributions()), and aborted if these were not 100%
equivalent.

After adding a test to test_real_projects, we have now verified that
this sanity check holds for all of the venvs we use as part of our
real_projects test. Since the sanity check has not triggered, we can
conclude that these methods are in fact equivalent, and we can drop
this sanity check.