Add venv setting and --venv command-line option
Hook it up to the venv_path argument to resolve_depenedencies(), to
allow the user to control which venv FawltyDeps uses to map dependencies
into import names.

For now we only accept a single venv dir here, but this might change in
the near future.

TODO: Write tests to verify that setting the command-line/config option
has the desired effect.