main: Write a friendly success message when we find no issues
When checking a project in the (default) summary mode, and no problems
were found, we would still print the following, confusing message to our
users:

  For a more verbose report re-run with the `--detailed` option.

Only print that message when there is indeed a preceding report.

Otherwise, when we're checking for undeclared/unused, and no problems
were found, print this friendlier message instead:

  No unused or undeclared dependencies detected.

Suggested by @koterpillar in #222.
This fixes #222.