README.md: Add documentation of --venv option
Also rewrite the last two FAQ questions in light of the new --venv
option. Use the second-last question as an opportunity to drill more
into the details that I feel are too verbose for the new section that
introduces --venv and the concept/importance of the Python environment
for mapping deendencies to import names via installed packages.

Co-authored-by: Nour El Mawass <nour.elmawass@tweag.io>