Refactor `extract_declared_dependencies` parser APIs in preparation for #227 (#229)
* test_extract_declared_dependencies_succes: Harmless test refactoring

Move the dependency_factory() call into the test case instead of in the
parametrized test data. Also remove unnecessary "__" prefix from test
ids.

* extract_declared_dependencies: Refactor API of parser functions

Take a single Path argument instead of separate contents + source. This
allows the parser function itself to choose how to open the file and
pass it to the underlying parser. This is also necessary for the
upcoming rewrite of our requirements.txt parser (#227).

* test_extract_declared_dependencies_pyproject_toml: Remove unnecessary dedent()

* extract_declared_dependencies: Rename parser functions

The rename reflects that these now parse file paths, and not extracted
file contents.

---------

Co-authored-by: Vince Reuter <vince.reuter@gmail.com>