main: Fix crash when encoding Settings.code items
Since we changed Settings.code from PathOrSpecial to Set[PathOrSpecial],
we now have to be able to represent a set that contains both Path
objects and the "<stdin>" special case. Simply passing this set to
sorted() causes a crash, since Path and str are not orderable (see #233
for details). Instead, sort the set using the _string representation_
of the items in the set.

Fixes #233.