packages.LocalPackageLookup: Accept more Python environments types
There are more types of Python environments than those that are covered
by the "virtualenv" (or "venv") definition:
- Environments prepared by e.g. poetry2nix, as shown in issue #236
- Environments that correspond to complete Python installations, like
  e.g. what you'd find under /usr or /usr/local.

We want to support all of these, and not only virtualenvs specifically.
Thus we cannot use the existence of a pyvenv.cfg file to identify a
Python environment. Instead we make our own definition:

  Any directory that contains a bin/python file, and also has a
  site-packages directory under lib/pythonX.Y (for some X.Y not
  necessarily the same as the current Python version) is to be
  considered a Python environment.

From this commit, we will happily use the lib/pythonX.Y/site-packages
subdirectory of any given directory that matches the above definition,
to resolve dependencies -> import names.