Rename --venv option to --pyenv
--pyenv better encapsulates that this can point to more than just
virtualenvs, including poetry2nix environment as well as "full" Python
installs as e.g. found in /usr or /usr/local on many distros.

Best to rename this _now_ before this new option makes it into a
released version of FawltyDeps.

Update documentation and tests in addition to the main code itself.