test_sample_project: Refactor to use BaseProject/BaseExperiment
This is a continuation of the previous commit: We now refactor
test_sample_project to adopt the common base classes introduced in the
previous commit. The new Experiment and SampleProject subclasses
accurately portrays what makes sample_projects different from
real_projects:

- Experiment adds .code and .deps members to allow customization of the
  Settings object that is created for each experiment. Experiment also
  has a .settings() method to create this Settings object for use in an
  experiment.
- SampleProject adds a .path member to point to the test/sample_projects
  subdir that contains this sample_project. It also implements the
  .collect() factory method to walk the subdirectories under
  test/sample_projects and create the corresponding objects.

This leaves the actual test function to simply parametrize over the
available projects/experiments, create an Analysis object with the
appropriate settings, and verify the contents of the analysis.

We add ample print() statements to help developers debug failing tests.

The existing sample_projects TOML files are modified slightly to
supported this refactoring: We rename the [analysis_result] section to
[experiment.<something>], and add a description as well as a
requirements list (empty in all current cases, as our existing
sample_projects need only rely on the identity mapping).

In sum, the objective of this refactoring is two-fold:
- Increase sharing of common code between sample_projects and
  real_projects.
- Teach sample_projects to be able to run multiple experiments per
  project, and have custom virtualenvs defined for each experiment.