sample_projects: Add empty project
This is a simple base case for test_sample_projects, but more than that:
we are about to refactor/develop the logic for traversing projects,
finding sources for imports and deps, as well as ignoring paths, and for
that we will add new tests that need sample project structures to
traverse. We will reuse the tests/sample_projects/* for this.

Also fix our property-based cmdline tests to handle projects with zero
files declaring dependencies.