Introduce traverse_project module
This module will take over the traversal of the project
files/directories specified in the Settings object, and generate sources
that will then be processed other parts of FawltyDeps, primarily
extract_imports and extract_declared_dependencies.

We start by reimplementing the file/directory traversal that currently
lives in the extract_imports and extract_declared_dependencies modules,
except that we focus on traversing the project _once_ and generating
CodeSource and DepsSource objects that will later be passed on to the
extract_imports and extract_declared_dependencies modules, respectively.

For now, the new code is only used by a test suite that exercises
various traversal scenarios on the projects under tests/sample_projects.