project_helpers.BaseProject: Clarify + document construction protocol
The construction of these Project/Experiment objects is too convoluted.
This commit tries to clarify some of the details.

We're working with inherited dataclasses here, and where we'd normally
solve this by customizing the constructors (e.g. by passing around a
TomlData object from subclass to base class, and have each level extract
the relevant details), we cannot do so with dataclasses.

Instead the instances are created by cls() calls in the respective
subclasses' .collect() methods, and the most appropriate thing we can do
in the base class is to provide a class method that extracts relevant
details from the TomlData, and returns them as a kwargs dict to be
passed to the subclass constructor.

We have already done so for the BaseExperiment class via the
._init_args_from_toml() class method, however the same pattern was not
as apparent in BaseProject.

This commit refactors the actual TOML _parsing_ into a parse_toml()
helper, and rephrases the rest into a ._init_args_from_toml() class
method on BaseProject that corresponds to the same from BaseExperiment.