test_cmdline: Unify testing of log messages
The _ordering_ of log messages is not something we should be too strict
about in tests, as it just makes it harder to refactor our main logic
without having to deal with lots of "collateral diff damage" in our
tests.

Establish a common pattern across all our command-line tests for
verifying expected log messages: We declare a list of expected log
message lines, we then split the stderr into lines and use the
assert_unordered_equivalence() helper to communicate that the _ordering_
of log messages is not important.