packages: Change BasePackageResolver API: resolve all names in one call
With the introduction of a resolver to temporarily install packages
remotely (from PyPI), we want to resolve all packages in one go instead
of resolving one package at a time (as this allows for a single `pip
install` command instead of one command per package).

Change the BasePackageResolver.lookup_package() method into a
BasePackageResolver.lookup_packages() method which converts a set of
package names into a dictionary mapping a subset of those names (i.e.
the names that could be successfully resolved by this resolver) into
their corresponding Package objects.

The resolve_dependencies() function is changed accordingly: Instead of
resolving one package at a time (iterating through the available
resolvers in order), we now instead loop over the available resolvers
_first_, passing all package names to the first resolver in the list,
and keeping track of which names was _not_ resolved by that resolver.
We then pass the unresolved names on to the next resolver in the list,
and so on, until either all names have been resolved (at which point we
don't need to consult any more resolvers), or we have reached the end of
the resolvers list.

Currently, the last resolver in the hardcoded list of resolvers is the
IdentityMapping, which guarantess that all the given package names will
get resolved. (This will change in the future.)