main: Refactor Analysis object to remove annoying asserts
Our Analysis object has long been annoying to type-check: Its members
are mostly optional, and whether or not they are set/calculated depends
on the actions that have been requested in the given Settings object.

This has led to a long and unwieldy .create() method, where we have
several if-blocks in sequence to figure out exactly which members are
required, and calculating them accordingly.

This is better(?) solved by implementing these members as "cached
properties", that is, methods that are called at most _once_ to
calculate their value, thenceforth cached. If a member is never
referenced, its value is never calculated. These semantics are much
closer to what we want for our Analysis object[^1].

Since these cached properties reference each other, we get an automatic
dependency relationship between the members which means that we no
longer have to maintain this dependency relationship manully inside
Analysis.create(). Instead, Analysis.create() only needs to make sure
that the requested members are realized/calculated, and the dependencies
between them will take care of the rest[^2].

On caveat worth mentioning: In our JSON output, we don't want to trigger
calculation of a member that was not required by the requested actions,
hence we here need direct access to the cached attribute _without_
causing it to be realized/calculated. This - in addition to having to
support Python v3.7 - is why we cannot use functools.cached_property
directly.

Hence, we instead have to provide our own @calculated_once decorator,
that, together with @property gets the job done.

Another casualty of this refactoring is that we are now adding so much
"magic" to the Analysis class that it no longer makes sense to define it
using @dataclass. Instead, we implement is as a regular class, and go
via an intermediate dictionary when we create the JSON output. This is
not optimal, but I cannot see a way to more seamlessly combine all the
things we want to do here...

[^1]: Although it does carry an extra cost: If one of these members is
unnecessarily referenced, it (and any other members it depends on) will
now be auto-calculated (i.e. costing runtime performance) instead of
raising a TypeError (when attempting to reference something inside the
None value).

[^2]: We already have tests that verify (via our JSON output) that we
don't calculate more Analysis member than we strictly need to, based on
the given actions.