noxfile.py: Install FD in editable mode in Nox sessions
This should help increase the usefulness of Nox's -R option (to reuse
virtualenvs from the last run). This option has been unreliable so far,
as it would end up re-testing the same old FD version that was installed
in the old virtualenv, instead of reflecting your latest changes to the
FD source tree.

With this, the FD installed in the Nox environments should now always
reflect the latest changes in the worktree.