extract_declared_dependencies: Move parser_choice into DepsSource
The parser choice was passed into both validate_deps_source() (to help
select which files become DepsSource objects), and into parse_source()
(to select the parser to be used for each DepsSource object). This is
redundant when we instead can record the parser_choice as part of the
DepsSource object.