test_cmdline: Spell out --verbose when we _do_ use it
For the remaining cases where we _do_ want/need to test INFO-level log
messages, spell out --verbose instead of the -v shorthand (which is
easy to overlook in a command line).

Also expand a couple of _ variables to _output, which I find more
consistent with the surrounding code.

This commit changes no behavior.