test_cmdline: Rename some tests
After we introduced the --detailed and --summery options
(which offloads some logic from --verbose/--quiet), some of the test
names have not been entirely precise in terms of what they intend to
test (or what options they pass on the command line). Fix that.

This commit does not change any behavior.