test_cmdline: Add TODO comment about potentially missing log message
This test name implies that we should have an INFO-level log message
when we traverse an empty dir looking for declared deps:

  test_list_deps__empty_dir__verbosely_logs_but_extracts_nothing

But the actual test asserts that no INFO-level logs are written.

I'm not sure which is the intended behavior, but the inconsistency
should be fixed one way or the other. For now, simply flag the issue.