test_cmdline_args_in_combination_with_config_file: Test logs + exit code
For this parametrized test, add verification that no >=WARNING-level log
messages are produces, and that the return code is within the expected
set (i.e. 0, 3, or 4, signalling a controlled/expected completion).

This goes somewhat in the opposite direction of the surrounding commits
that _remove_ log message verification from test_cmdline. However, in
this case, I was not able to convert run_fawltydeps_subprocess() into
run_fawltydeps_function() without breaking the test, and once we're
calling the subprocess version, we should at least _consider_ whether it
is relevant to test the log messages and return code in this test:

This test is an end-to-end test that exercises the cascade/selective
overriding of settings between the config file and the command line, and
it is somewhat relevant (IMHO) to ensure that the tested combinations
produce no WARNING- or ERROR-level logs, and that the return code is
within the expected set.