Add fawltydeps.__main__ module to allow 'python -m fawltydeps'
It is common for Python packages that offer a command-line interface to
expose this via the __main__ module in the package. See [1] for info.

Also add some tests to verify the various ways we can now invoke the
fawltydeps command-line interface:

- using the `fawltydeps` entry point
- calling the fawltydeps.main.main() function from inside Python
- running `python -m fawltydeps`

These tests discovered that the program name shown at the start of the
--help message did not make sense when using the latter two invocations.
Fix that by explicitly passing the program name to argparse.

Fixes: https://github.com/tweag/FawltyDeps/issues/309

Suggested-by: Timothée Mazzucotelli <pawamoy@pm.me>

[1]: https://docs.python.org/3/library/__main__.html#main-py-in-python-packages