packages: Support PEP582 __pypackages__ directories for --pyenv
Even though PEP582[1] was recently rejected, it still has a significant
user base via tools like PDM (see issue #307), and it's worth supporting
these as long as the added burden is not great.

The layout of __pypackages__ directories is different from what we
currently use to identify a Python environment: Where we expect to find:

  .../bin/python
  .../lib/pythonX.Y/site-packages/<packages installed here>

there is instead:

  __pypackages__/X.Y/bin/ (but NO python)
  __pypackages__/X.Y/lib/<packages installed here>

Since our LocalPackageResolver.pyenv_path member already points directly
to the directory holding the packages (the site-packages dir in the
former case), we can fairly easily add support for __pypackage__
directories, simply by providing an alternative check for the
__pypackages__ directory structure.

This should provide a fix/workaround for issue #307 by allowing

  --pyenv path/to/__pypackages__

to be passed on the command-line. FawltyDeps should then be able to find
the appropriate deps, and correctly match imports and declared deps as a
result.

Note that we still require the venv/__pypackage__ dir to be passed with
--pyenv; auto-detection of environments (both kinds) is still a planned
feature.

[1]: https://peps.python.org/pep-0582/