packages: Refactor Package to add more debugging information
Until now we have designed the Package object to be able to accept
package-name-to-import-names mappings from _multiple_ resolvers
(signified with the DependenciesMapping enum). However, our current
design relies on each Package object being built by a _single_ resolver,
so the multiple mappings are never used in practice.

What would be nice, however, is to be able to attach some more debug
information to each Package, specifically pointing to where we found
the relevant package-name-to-import-names mappings represented in this
object. (This will soon become more important when we add support for
multiple --pyenv options.)

In this refactoring we reshape the Package object as follows:
- We remove the DependenciesMapping enum. It has served as a simple
  proxy for the resolver class/type used to create the Package object.
- Instead, we record the responsible class itself directly into the
  Package object, by adding a .resolved_with member of type
  Type[BasePackageResolver]. When serializing Package objects to JSON,
  this member is serialized to a simple string: the class name.
- We remove the .mappings member (which only ever has a single member in
  practice)
- Instead, we add a .debug_info member which each resolver is free to
  populate with any information (but defaults to None). The only
  real requirement is that the information be JSON-serializable.
  For now, we use these kinds of debug information:
    - UserDefinedMapping adds nothing (None) for now, but this will soon
      change to indicate from where the relevant user-defined mapping
      was read
    - LocalPackageResolver stores a dictionary mapping the site-packages
      path where a Package was found to the import names found in that
      install location. This will soon become even more useful when we
      add support for multiple --pyenv arguments, which will potentially
      merge Package definitions from multiple site-packages directories.
    - TemporaryPipInstallResolver overrides the debug_info from the
      underlying LocalPackageResolver to a simple string indicating that
      packages were temporarily installed with `pip install`.
    - IdentityMapping uses None as it has no extra information to give.

Most of the diff is occupied by associated test updates to deal with the
reshaped Package objects. Except for a very few unit tests, we don't
bother verifying the exact contents of the .debug_info member, instead
we liberally use a ignore_package_debug_info() helper to remove
.debug_info details before comparing Package objects.