LocalPackageResolver: Resolve packages using _first_ match in sys.path
When LocalPackageResolver is used to look up packages in the _current_
Python environment, it should take care to use the same package that
Python itself would end up using when importing a package of that name.

Until now, the way we have used importlib_metadata (specifically the way
we iterate over MetadataPathFinder().find_distributions()), if a
packages occurs multiple times in sys.path, we would end up using the
_last_ instance, whereas Python would use the _first_. Fix our iteration
to be in line with Python.