settings.print_toml_config(): Represent sets as sorted lists
In our --json output, we customize pydantic's JSON encoder to represent
sets as sorted lists (thus yielding stable output). Do the same for when
we generate TOML config output (which goes via a JSON encoding step, and
thus is subject to the same issue with how sets are encoded).

This was discovered while attempting to test how multiple 'pyenvs' were
represented in our TOML output.