sample_projects: Add pyenv_galore 'project'
This is not a real project, but merely a collection of different kinds
of Python environments that we should support.

In a future commit, the project traversal code in FawltyDeps will learn
to auto-detect Python environments, and ignore code/deps sources
otherwise found within these. When that is in place, this sample
project's expected.toml will become considerably simpler, as everything
should be correctly detected with default settings. For now, though, we
need to pass very explicit code/deps/pyenvs options in order to
compensate for FawltyDeps' lack of smarts.