test_sample_projects: Accept new 'pyenvs' directive in expected.toml
Until now, all sample experiments have been run with a cached venv that
is populated with the packages listed in the 'requirements' directive
(empty by default).

This commit allows experiments to declare a `pyenvs` list of directories
(relative to the sample project) to be passed (via settings.pyenvs)
instead of the cached venv. This will allow sample projects to embed
Python environments within their project, which can then be used e.g.
to test the traversal of the project directories.

When given, the `pyenvs` directive will _override_ the `requirements`
directive, meaning that the cached venv will no longer be used to
resolve packages, only the paths passed to 'pyenvs' will be used.