main: Use PyEnvSource objects when resolving dependencies
This starts requesting PyEnvSource objects when we call
traverse_project.find_sources(), and pass the paths from these object on
to resolve_dependencies().

In other words, this commit is when we first start actually using the
new traversal logic for Python environments in FawltyDeps.

On the test side, we can finally remove the overly explicit
specification needed to correctly traverse/parse the pyenv_galore sample
project. Instead, this project is now successfully handled with default
settings:
 - pyenvs set to the project root is able to auto-detect all the Python
   environments found within.
 - As a result, there is no code, nor any deps files found (since the
   project consists of Python environments only).