test_traverse_project: Prepare for testing with 'pyenvs'
We are about to introduce Python environments into the traverse_project
logic. When testing the new logic we will - in addition to passing in
code + deps via settings and expecting corresponding CodeSource and
DepsSource objects to be created - also do the same for settings.pyenvs,
resulting in the creation of PyEnvSource objects.

The default mode of operation will be to auto-detect Python environments
under the current project dir (as we currently do for code + deps).

In order for the new logic not to interfere with the existing code/deps
tests, we must therefore explicitly pass an empty pyenvs set to the
existing tests. This commit does exactly that. A lot of diff noise to
ensure that a future change does not change how these existing tests
will work.