traverse_project: Add support for finding/validating Python environments
In addition to finding/creating CodeSource and DepsSource objects in
traverse_project.find_sources(), we now also add support for PyEnvSource
objects. Each of these objects encapsulate a single directory within a
Python environment where we expect to find Python packages installed.

In addition to validating the settings.pyenvs paths (similar to what the
LocalPackageResolver constructor already does), this also adds the logic
to find new Python environments within the directories being traversed.

That is, until now we have only supported entries in settings.pyenvs
pointing directly to (a directory within) a Python environment, but with
this commit we can also start finding Python environments beneath a
given directory.

In others words: given a project containing a ".venv" subdir, we
previously needed 'pyenvs' to point directly at ".venv" in order to find
this Python environment, but with this commit we can soon support
'pyenvs' pointing at the project root and still be able to automatically
find ".venv" inside the project.

Nothing of this is user-visible yet, as we still require the high-level
business logic in fawltydeps.main to actually request, and use the new
PyEnvSource objects created in this commit.

Based on the recent additions to test_sample_projects and
sample_projects, we add a bunch of tests to verify that we are able to
both precisely specify which Python environments should be included in
our analysis, as well as being able to automatically find and use many
different Python environments given a more general specification.