LocalPackageResolver: Construct from PyEnvSource objects
Require Python environment paths to have been converted into PyEnvSource
objects _before_ calling the LocalPackageResolver constructor and the
setup_resolvers wrapper.

Provide a pyenv_sources() helper function to do the conversion for call
sites where this is necessary.

This removes an ugly duplication that was introduced when we moved pyenv
logic into traverse_projects: We would call
LocalPackageResolver.find_package_dirs() for each pyenv path in
validate_pyenv_source() to create the PyEnvSource objects, and then, in
the LocalPackageResolver constructor we would call
LocalPackageResolver.find_package_dirs() _again_ for each and every
path passed into it. The PyEnvSource objects would in effect be
validated _twice_, the second time being an expensive no-op.

With this commit, we _require_ PyEnvSource objects to be passed directly
to LocalPackageResolver, and their validation now happens only _once_
(with the call to LocalPackageResolver.find_package_dirs() from
pyenv_sources()).